vst2: split up adapter into a few smaller files
